Fix wrong size for guild server icon
When there is an issue with a discord guild server the icon it shows is the incorrect size.  It gets sized 50px x 50px which is the normal size for Discord.  Checking inspector shows that ".guilds-wrapper .guilds-error" is overriding the theme sizing.  Changing this line to match allows it to get proper sizing and line height so that it looks correct with all the other guild icons.